make VP compile with clang, also fix bugs in the audio options dialog
32bit build has probs with linking though, due to DX dinput, 64bit build works

git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3844 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ dialogs/EditorOptionsDialog.cpp
@@ -264,7 +264,7 @@ void EditorOptionsDialog::OnOK()
     size_t units = SendMessage(hwndUnits, CB_GETCURSEL, 0, 0);
     if (units == LB_ERR)
         units = 0;
-    SaveValueInt("Editor", "Units", units);
+    SaveValueInt("Editor", "Units", (int)units);
 
     CDialog::OnOK();
 }
